added goal form to popup
