+++ just-move/src/components/UserHomePage/UserHomePage.js
@@ -1,16 +1,12 @@
 import { useState, useRef } from 'react';
 import React from 'react';
 import { Goal } from "./Goal";
-import Modal from "./Modal";
-import { nanoid } from 'nanoid';
-import IconButton from 'rsuite/IconButton';
-import PlusIcon from '@rsuite/icons/Plus';
+import PopupGoalForm from "./PopupGoalForm";
+import FadeIn from 'react-fade-in';
 
 export function UserHomePage () {
 
   const goalRef = useRef(null); 
-  const intrinsicRef = useRef(null);
-  const extrinsicRef = useRef(null);
 
   const [goals, setGoals] = useState([]);
 
@@ -22,34 +18,6 @@ export function UserHomePage () {
   });
 
   const [popupBtn, setPopupBtn] = useState(false);
-
-  const handleGoalsChange = (e) => {
-    e.preventDefault();
-    const goalName = e.target.getAttribute("name");
-    const goalValue = e.target.value;
-    const newGoalData = { ...addGoalData };
-    newGoalData[goalName] = goalValue;
-    setGoalData(newGoalData);
-  }
-
-  const handleAddNewGoal = (e) => {
-    e.preventDefault();
-
-    const newGoal = {
-      id: nanoid(),
-      goal: addGoalData.goal,
-      intrinsicMotivation: addGoalData.intrinsicMotivation,
-      extrinsicMotivation: addGoalData.extrinsicMotivation,
-      progress: " ",
-    }
-
-    const newGoals = [...goals, newGoal];
-    setGoals(newGoals);
-
-    goalRef.current.value = "";
-    intrinsicRef.current.value = "";
-    extrinsicRef.current.value = "";
-  }
 
   const handleDeleteGoal = (goalId) => {
     const newGoals = [...goals];
@@ -63,56 +31,18 @@ export function UserHomePage () {
 
   return (
     <div>
-      <h1>Welcome to your home page!</h1>
+      <h1>Just Move</h1>
       <button onClick={() => setPopupBtn(true)}>
-        Open
+        Add a new goal!
       </button>
-      <Modal trigger={popupBtn}
-      setPopupBtnTrigger={setPopupBtn}>
-      </Modal>
-      <form onSubmit={handleAddNewGoal}>
-        <h2>Let's Create a Goal!</h2>
-        <div className="form-group">
-          <label>
-            Goal:
-          </label>
-          <input
-            type="input"
-            name="goal"
-            placeholder="Enter a goal..."
-            ref={goalRef}
-            className="form-control"
-            onChange={handleGoalsChange}
-          />
-        </div>
-        <div className="form-group">
-          <label>
-            Intrinsic Goal:
-          </label>
-          <input
-            type="input"
-            name="intrinsicMotivation"
-            placeholder="Enter your intrinsic motivation for this goal..."
-            ref={intrinsicRef}
-            className="form-control"
-            onChange={handleGoalsChange}
-          />
-        </div>
-        <div className="form-group">
-          <label>
-            Extrinsic Goal:
-          </label>
-          <input
-            type="input"
-            name="extrinsicMotivation"
-            placeholder="Enter your extrinsic motivation for this goal..."
-            ref={extrinsicRef}
-            className="form-control"
-            onChange={handleGoalsChange}
-          />
-        </div>
-        <IconButton type="submit" icon={<PlusIcon />} appearance="primary" color="green"></IconButton>
-      </form>
+      <PopupGoalForm trigger={popupBtn}
+      setPopupBtnTrigger={setPopupBtn}
+      goalRef={goalRef}
+      addGoalData={addGoalData}
+      setGoalData={setGoalData}
+      goals={goals}
+      setGoals={setGoals}>
+      </PopupGoalForm>
       <table id="goals-table" className="table mt-5">
       <thead>
         <tr>
